+++ core/src/main/scala/com/stratio/sparta/core/models/WorkflowValidationMessage.scala
@@ -7,10 +7,8 @@
 package com.stratio.sparta.core.models
 
 case class WorkflowValidationMessage(message: String, step: Option[String] = None, subStep: Option[String] = None) {
-
-  override def toString: String = if(step.isDefined){
-    s"$message in step ${step.get}"
-  } else message
+  
+  override def toString: String = s"$message${step.map(s => s" in Step [$s]").getOrElse("")}"
 
 }